add: pitest for multi-module projects
